requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncgui==0.6.1
asynckivy==0.6.2
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
darkdetect==0.8.0
distro==1.9.0
dnspython==2.6.1
docutils==0.21.2
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
materialyoucolor==2.0.9
openai==1.31.1
packaging==24.0
pillow==10.3.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pymongo==4.7.3
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tk==0.1.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
